BE-820 Create a function to replace path.Join for URLs
